refactor: move eth2wrap to a new crate + make CachedValidatorsProvider async
