+++ Cargo.toml
@@ -6,11 +6,13 @@ members = [
     "crates/cli",
     "crates/cluster",
     "crates/core",
+    "crates/core-utils",
     "crates/crypto",
     "crates/dkg",
     "crates/eth2api",
     "crates/eth2util",
     "crates/eth1wrap",
+    "crates/eth2wrap",
     "crates/featureset",
     "crates/k1util",
     "crates/relay-server",
@@ -131,6 +133,8 @@ pluto-tracing = { path = "crates/tracing" }
 pluto-p2p = { path = "crates/p2p" }
 pluto-peerinfo = { path = "crates/peerinfo" }
 pluto-frost = { path = "crates/frost" }
+pluto-eth2wrap = { path = "crates/eth2wrap" }
+pluto-core-utils = { path = "crates/core-utils"}
 
 [workspace.lints.rust]
 missing_docs = "deny"
