+++ crates/app/Cargo.toml
@@ -9,13 +9,11 @@ publish.workspace = true
 [dependencies]
 backon.workspace = true
 chrono.workspace = true
-pluto-core.workspace = true
 pluto-eth2api.workspace = true
 tokio.workspace = true
 tokio-util.workspace = true
 prost.workspace = true
 prost-types.workspace = true
-regex.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
 url.workspace = true
