--- crates/app/src/eth2wrap/mod.rs
@@ -1,2 +1,0 @@
-/// Validate Beacon node versions
-pub mod version;
