+++ crates/core-utils/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "pluto-core-utils"
+version.workspace = true
+edition.workspace = true
+repository.workspace = true
+license.workspace = true
+publish.workspace = true
+
+[dependencies]
+chrono.workspace = true
+regex.workspace = true
+thiserror.workspace = true
+tracing.workspace = true
+pluto-crypto.workspace = true
+serde.workspace = true
+hex.workspace = true
+
+[build-dependencies]
+built.workspace = true
+
+[lints]
+workspace = true
