+++ crates/core-utils/src/lib.rs
@@ -1,0 +1,7 @@
+//! Dependency-free core utilies
+
+/// Semver version parsing utilities.
+pub mod version;
+
+/// PubKey type
+pub mod pubkey;
