+++ crates/core/Cargo.toml
@@ -20,11 +20,12 @@ k256.workspace = true
 libp2p.workspace = true
 vise.workspace = true
 pluto-crypto.workspace = true
+pluto-core-utils.workspace = true
 pluto-eth2api.workspace = true
+pluto-eth2wrap.workspace = true
 pluto-k1util.workspace = true
 prost.workspace = true
 prost-types.workspace = true
-regex.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 base64.workspace = true
