+++ crates/eth2wrap/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "pluto-eth2wrap"
+version.workspace = true
+edition.workspace = true
+repository.workspace = true
+license.workspace = true
+publish.workspace = true
+
+[dependencies]
+async-trait.workspace = true
+pluto-eth2api.workspace = true
+pluto-core-utils.workspace = true
+tokio.workspace = true
+regex.workspace = true
+thiserror.workspace = true
+tracing.workspace = true
+
+[build-dependencies]
+pluto-build-proto.workspace = true
+
+[dev-dependencies]
+pluto-testutil.workspace = true
+wiremock.workspace = true
+test-case.workspace = true
+
+[lints]
+workspace = true
