+++ crates/eth2wrap/src/lib.rs
@@ -1,0 +1,7 @@
+//! Ethereum CL RPC client management.
+
+/// Validate Beacon node versions
+pub mod version;
+
+/// Cache of validators retrieved from the Beacon node.
+pub mod valcache;
